Gave PlaceInitialRoadCommand it's own file and named it properly
