+++ sod-server/src/commands/base/PlaceRoadCommand.ts
@@ -2,18 +2,9 @@ import { Command } from '@colyseus/command'
 import { MyRoom } from '../../rooms/MyRoom'
 import { CardNames } from '../../rooms/schema/Card'
 
-interface Payload {
+export interface Payload {
   edgeId: string
   playerId: string
-}
-export class PlaceInitRoadCommand extends Command<MyRoom, Payload> {
-  execute(payload: Payload) {
-    const player = this.room.state.players.get(payload.playerId)
-    const availableRoad = player.roads.find((x) => !x.edge)
-    if (availableRoad) {
-      availableRoad.edge = payload.edgeId
-    }
-  }
 }
 
 export class PlaceRoadCommand extends Command<MyRoom, Payload> {
