+++ sod-server/src/stateMachines/events/base.ts
@@ -1,8 +1,8 @@
-import { PlaceInitRoadCommand } from '../../commands/base/PlaceRoadCommand'
+import { PlaceInitialRoadCommand } from '../../commands/base/PlaceInitialRoadCommand'
 import { PlaceSettlementCommand } from '../../commands/base/PlaceSettlementCommand'
 
 type PlaceSettlementEvent = { type: 'PLACE_SETTLEMENT'; payload: PlaceSettlementCommand['payload'] }
-type PlaceRoadEvent = { type: 'PLACE_ROAD'; payload: PlaceInitRoadCommand['payload'] }
+type PlaceRoadEvent = { type: 'PLACE_ROAD'; payload: PlaceInitialRoadCommand['payload'] }
 type RollDiceEvent = { type: 'ROLL_DICE' }
 type EndTurn = { type: 'END_TURN' }
